+++ lib/kitchen/directions/bake_index/v1.rb
@@ -119,26 +119,38 @@ module Kitchen::Directions::BakeIndex
       @index = Index.new
 
       book.pages.terms.each do |term_element|
-        # Markup the term
         page = term_element.ancestor(:page)
         term_element.id = "auto_#{page.id}_term#{term_element.count_in(:book)}"
+        page_title = page.title.text
+        add_term_to_index(term_element, page_title)
+      end
 
-        group_by = term_element.text.strip[0]
-        group_by = I18n.t(:eob_index_symbols_group) unless group_by.match?(/\w/)
-        term_element['group-by'] = group_by
-
-        # Add it to our index object
-        @index.add_term(
-          Term.new(
-            text: term_element.text,
-            id: term_element.id,
-            group_by: group_by,
-            page_title: page.title.text.gsub(/\n/, '')
-          )
-        )
+      book.chapters.composite_pages.terms.each do |term_element|
+        page = term_element.ancestor(:composite_page)
+        chapter = term_element.ancestor(:chapter)
+        term_element.id = "auto_composite_page_term#{term_element.count_in(:book)}"
+        chapter_number = chapter.count_in(:book)
+        page_title = "#{chapter_number} #{page.title.text.strip}".strip
+        add_term_to_index(term_element, page_title)
       end
 
       book.first('body').append(child: render(file: 'v1.xhtml.erb'))
+    end
+
+    def add_term_to_index(term_element, page_title)
+      group_by = term_element.text.strip[0]
+      group_by = I18n.t(:eob_index_symbols_group) unless group_by.match?(/\w/)
+      term_element['group-by'] = group_by
+
+      # Add it to our index object
+      @index.add_term(
+        Term.new(
+          text: term_element.text,
+          id: term_element.id,
+          group_by: group_by,
+          page_title: page_title.gsub(/\n/, '')
+        )
+      )
     end
 
   end