+++ spec/directions/bake_index/v1_spec.rb
@@ -35,6 +35,20 @@ RSpec.describe Kitchen::Directions::BakeIndex::V1 do
             <span data-type="term">ΔE</span>
             <span data-type="term"><em>sp</em><sup>3</sup><em>d</em><sup>2</sup> orbitals</span>
           </div>
+          <div data-type="composite-chapter">
+            <div data-type="document-title">Chapter Review</div>
+            <div data-type="composite-page">
+              <div data-type="title">EOC Section Title</div>
+              <span data-type="term">composite page in a composite chapter</span>
+            </div>
+          </div>
+        </div>
+        <div data-type="chapter">
+          <div data-type="page" id="p4"/>
+          <div data-type="composite-page">
+            <div data-type="document-title">Another EOC Section</div>
+            <span data-type="term">composite page at the top level</span>
+          </div>
         </div>
       HTML
     )
@@ -63,6 +77,21 @@ RSpec.describe Kitchen::Directions::BakeIndex::V1 do
               <span class="os-term" group-by="Symbols">&#x394;E</span>
               <a class="os-term-section-link" href="#auto_p2_term4">
                 <span class="os-term-section">1.1 First Page</span>
+              </a>
+            </div>
+          </div>
+          <div class="group-by">
+            <span class="group-label">C</span>
+            <div class="os-index-item">
+              <span class="os-term" group-by="c">composite page at the top level</span>
+              <a class="os-term-section-link" href="#auto_composite_page_term2">
+                <span class="os-term-section">2 Another EOC Section</span>
+              </a>
+            </div>
+            <div class="os-index-item">
+              <span class="os-term" group-by="c">composite page in a composite chapter</span>
+              <a class="os-term-section-link" href="#auto_composite_page_term1">
+                <span class="os-term-section">1 EOC Section Title</span>
               </a>
             </div>
           </div>